+++ src/game.hpp
@@ -1,22 +1,26 @@
 #ifndef _LoX_GAME
 #define _LoX_GAME
 
+#include "state.hpp"
+#include "item.hpp"
 #include "level.hpp"
-#include "renderer.hpp"
 #include "character.hpp"
 
-class Game {
+class Game : public State {
     std::vector<Level> _dungeon;        // array of dungeons
+    uint _level;                        // current dungeon level
     Character _player;                  // player character
     std::vector<item_ptr> _inventory;   // player inventory (TODO: Move to character class)
-    Renderer _renderer;                 // game renderer
 
 public:
     Game();
     ~Game();
 
-    /* method called to start the game, returns exit status */
-    int Execute();
+    /* method called to update the game, overrides from base State class */
+    void Update();
+
+    /* method called to draw the screen, overrides from base State class */
+    void Draw();
 };
 
 #endif