+++ src/state.cpp
@@ -1,0 +1,4 @@
+#include "state.hpp"
+
+uchar State::_state;
+Renderer State::_ren;